+++ src/extension.ts
@@ -4,6 +4,7 @@ import { initTaskCmd } from './commands/tasks_commands';
 import { taskViewDataProvider} from './view/tasks_view';
 import { initConfigView } from './view/config_view';
 import { initConfigManager } from './utils/config_manager';
+import { initTaskManager } from './model/tasks_model';
 
 export var extensionRootUri: vscode.Uri;
 export var userName: string;
@@ -12,8 +13,9 @@ export function activate(context: vscode.ExtensionContext) {
 	console.log('Congratulations, your extension "slurm--" is now active!');
 
 	extensionRootUri = vscode.Uri.file(context.extensionPath);
-
+	
 	initConfigManager(context);
+	initTaskManager(context);
 
 	initTaskCmd(context);