+++ src/model/tasks_model.ts
@@ -1,8 +1,11 @@
 import * as vscode from 'vscode';
+import * as fs from 'fs';
+import * as path from 'path';
 
 export enum TaskState {
     R = "RUNNING",
-    PD = "PENDING"
+    PD = "PENDING",
+    CG = "COMPLETING",
 }
 
 
@@ -31,7 +34,7 @@ export class FilePath {
     constructor(path: string) {
         this.uri = vscode.Uri.file(path);
         let uriPart = path.split('/');
-        this.name = uriPart[uriPart.length-1];
+        this.name = uriPart[uriPart.length - 1];
     }
 
     toString() {
@@ -58,8 +61,10 @@ export class Task {
     readonly out_path: FilePath;
     readonly err_path: FilePath;
     reason: string;
+    finished: boolean;
     //JobID,Name:255,Username:20,State:20,NodeList,Gres:50,TimeLimit,TimeUsed,Command:255,STDOUT:255,STDERR:255,Reason:100
     constructor(jobid: string, name: string, user: string, state: string, node: string, gres: string, limit_time: string, runing_time: string, command: string, out_path: string, err_path: string, reason: string)
+    constructor(jobid: number, name: string, user: string, state: string, node: string, gres: Gres, limit_time: string, runing_time: string, command: string, out_path: FilePath, err_path: FilePath, reason: string, finished: boolean)
 
     constructor(
         jobid: number | string,
@@ -71,9 +76,10 @@ export class Task {
         limit_time: string,
         runing_time: string,
         command: string,
-        out_path: string,
-        err_path: string,
-        reason: string
+        out_path: string | FilePath,
+        err_path: string | FilePath,
+        reason: string,
+        finished?: boolean,
     ) {
         this.jobid = typeof jobid === "string" ? parseInt(jobid) : jobid;
         this.name = name;
@@ -84,9 +90,28 @@ export class Task {
         this.limit_time = limit_time;
         this.runing_time = runing_time;
         this.command = command;
-        this.out_path = new FilePath(out_path.replace('%j', jobid.toString()));
-        this.err_path = new FilePath(err_path.replace('%j', jobid.toString()));
+        this.out_path = typeof out_path === "string" ? new FilePath(out_path.replace('%j', jobid.toString())) : out_path;
+        this.err_path = typeof err_path === "string" ? new FilePath(err_path.replace('%j', jobid.toString())) : err_path;
         this.reason = reason === 'None' ? '' : reason;
+        this.finished = finished??false;
+    }
+
+    public static fromObject(obj: Task): Task {
+        return new Task(
+            obj.jobid,
+            obj.name,
+            obj.user,
+            obj.state,
+            obj.node,
+            obj.gres,
+            obj.limit_time,
+            obj.runing_time,
+            obj.command,
+            obj.out_path,
+            obj.err_path,
+            obj.reason,
+            obj.finished
+        );
     }
 
     public update(task: Task) {
@@ -95,26 +120,62 @@ export class Task {
         this.reason = task.reason;
     }
 
+    public finish() {
+        this.finished = true;
+    }
+
 }
 
 export class TaskManager {
-    private static _instance: TaskManager | null = null;
-    private constructor() { }
+    private storagePath;
+    private taskMap!: Map<number, Task>;
+    private initing: boolean = true;
 
-    static getInstance() {
+    private static _instance: TaskManager | null = null;
+    private constructor(context: vscode.ExtensionContext) {
+        this.storagePath = path.join(context.globalStorageUri.fsPath, 'tasks.json');
+        const dir = path.dirname(this.storagePath);
+        fs.mkdirSync(dir, { recursive: true });
+        this.load_task();
+        // console.log(this.storagePath)
+    }
+
+    static getInstance(context?: vscode.ExtensionContext): TaskManager {
         if (TaskManager._instance === null) {
-            TaskManager._instance = new TaskManager();
+            if (context === undefined) {
+                throw new Error(`init ${this.name} failed.`);
+            }
+            TaskManager._instance = new TaskManager(context);
         }
         return TaskManager._instance;
     }
 
-    private taskMap: Map<number, Task> = new Map();
+    private load_task() {
+        if (fs.existsSync(this.storagePath)) {
+            const jsonData = fs.readFileSync(this.storagePath, 'utf8');
+            const saveMap = JSON.parse(jsonData);
+            this.taskMap = new Map(saveMap);
+            this.taskMap.forEach((v, k) => this.taskMap.set(k, Task.fromObject(v)));
+            // console.log(this.taskMap)
+            // console.log(typeof saveMap)
+        } else {
+            this.taskMap = new Map();
+        }
+    }
+
+    private save_task() {
+        const arrData = Array.from(this.taskMap.entries());
+        const jsonData = JSON.stringify(arrData);
+        fs.writeFile(this.storagePath, jsonData, 'utf8', () => { });
+        // console.log('save!')
+    }
 
     public addTask(...tasks: Task[]) {
         tasks.forEach(value => this.taskMap.set(value.jobid, value));
     }
 
     public updateTask(...tasks: Task[]) {
+        // console.log(this.taskMap)
         const newId = tasks.map(value => value.jobid);
         const oldId = [...this.taskMap.keys()];
 
@@ -122,10 +183,13 @@ export class TaskManager {
         const addId = newId.filter(value => !updateId.includes(value));
         const dropId = oldId.filter(value => !updateId.includes(value));
 
-        dropId.forEach(value => this.taskMap.delete(value));
+
         this.addTask(...tasks.filter(value => addId.includes(value.jobid)));
         tasks.filter(value => updateId.includes(value.jobid))
-        .forEach(value=> this.taskMap.get(value.jobid)?.update(value));
+            .forEach(value => this.taskMap.get(value.jobid)?.update(value));
+        dropId.forEach(v => this.taskMap.get(v)?.finish());
+
+        this.save_task();
     }
 
     public deleteTask(...tasks: number[]): void;
@@ -141,13 +205,14 @@ export class TaskManager {
         }
     }
 
-    public getTask(name?: string): Task[] {
-        if (name === undefined) {
-            return [...this.taskMap.values()];
-        } else {
-            return [...this.taskMap.values()].filter((value) => { value.name === name; });
-        }
+    public getTask(): Task[] {
+        return [...this.taskMap.values()];
     }
+
 }
 
-export const taskManager = TaskManager.getInstance();
+export let taskManager: TaskManager;
+
+export function initTaskManager(context: vscode.ExtensionContext) {
+    taskManager = TaskManager.getInstance(context);
+}